Constraints on ListVariadics
Summary:
This diff brings support for constraints on ListVariadics. The constraints that can be specified are the same as for TypeVar, that is, Bounds and Explicits. Although ListVariadics currently could infer the bounds based on other values of the concatenation, this approach is too limited.

While Pyre had already internal support to express that a ListVariadic has constraints, all the logic related with it was empty, since it was not allow to directly specify the constraint of a ListVariadic. For that purpose, this diff defines the required logic, mostly the propagation of constraints to ensure that violation of constraints are identified, as well as subtyping between the various combinations of variadics.

Since subtyping of variables can be ambiguous, we support every case except when given something reduced to `[...,A,Ts1] <: [Ts2,B,...]`.

Finally, since parse_declaration does not have access to create_logic, currently only primitive types can be used as constraints, for that purpose a follow up diff is proposed. See comment:  https://www.internalfb.com/intern/diff/D23342739/?dest_fbid=606703123545772&transaction_id=343730890366364

Differential Revision: D23342739

fbshipit-source-id: 8b6848fdfcfc6d29be5be0970d1b2a58ad24b13e