treat `Top`-containing annotations as `Top` for error purposes
Summary: Any ast.Call nodes seem to be resolved as type `Top`. Previously, we explicitly check whether an annotation resolves as `Top` and treat this as a special case for error purposes; with this commit, we check whether the annotation contains anything that resolves to `Top`, and emit an error if so.

Reviewed By: shannonzhu

Differential Revision: D23887756

fbshipit-source-id: 34550fadde9fd39428d26a7a1351ae8b61d9ae9c