modify partial sink syntax to make impossible states unrepresentable
Summary:
Problem: Combined source rules are *weird*. With regular rules, we're totally fine with propagating a sink upwards, no matter which rule it corresponds to. However, if you consider multiple-source rules, the state is a bit different. If we have 2 different rules accepting the same sink with different source labels, when we trigger the other sink, we lose all context of where the original sink came from.

This manifested itself in practice when someone wrote a rule, and we started getting a lot of false positives. There's a couple of ways to solve this problem, but the approach I chose was:

1. Stop requiring the declaration of labels and sinks in the "sinks" bit.
2. Create a new "partial_sink" field for combined source rules. The name here must be unique, and the declaration of the name and labels is implicit.
3. Everything else is unchanged.

An alternative approach would be to propagate what source kinds matched the other sink when triggering. I decided against this because adding all the extra metadata adds a lot of complexity to our matching which we can design away if we restrict partial sinks to be unique.

Some benefits of this approach: The redundancy of re-declaring sinks is gone, the name of the sink should make it more obvious which combined source rule you're corresponding to (because these flows are hard to read). Also, we no longer have the confusion of people trying to specify `x: TaintSink[PartialSinkWithoutLabels]`.

The main con is that we're breaking the mental model a bit. This is precisely why we're using the name of "partial_sink" in the configuration: The idea is to lead people to the documentation if they don't know how to use these and make sure they're not confused with regular sinks.

Reviewed By: fahndrich

Differential Revision: D23940466

fbshipit-source-id: 2188103412c0c2eb89c05c691e12b115c320ee99