Allow source subtraces and use them in conjoined rules
Summary:
We have rules where we look for 2 sources reaching say Ent::genEnforce($vc, $id) in such a way that $vc is special (omni or PAT) and $id is user controlled.

Until now, we only emitted one trace (either to the VC or the the ID) and the other trace was hidden.

With this change the second trace should be visible as well as a subtrace.

Differential Revision: D23910481

fbshipit-source-id: 9067d17b532c74918f5fc9a27b22dcd78c96044e