Add no commit option to upgrade global version
Summary:
Some users are bending over backwards to shell out to version control to undo this commit.

This is supported in most upgrade commands, let's do it for upgrade global command to.

Note: upgrade commands could use some refactoring, it might be nice to pull this flag and some others higher up in the inheritance. Not going to put time into it right away though.

Reviewed By: pradeep90

Differential Revision: D24002273

fbshipit-source-id: 34b1be88796ebe17709c706e5ed180b992d52fa9