resolve taint for attribute accesses in unions properly
Summary:
Turns out that we were missing the taint for attributes in unions in attribute
accesses - let's recurse down into each possible element. Note that this is
only a problem in the forward analysis since in the backwards analysis we don't
have any attribute sinks/tito depending on the class, and simply rely on the
name.

Reviewed By: dkgi

Differential Revision: D24000600

fbshipit-source-id: 4d785605fb36b4027b5b19fb5bbb55cc75777228