start treating SkipOverrides-only models as non-existent
Summary:
Problem: We use the model syntax to mark functions as `SkipOverrides`. This has
the unfortunate effect of making models no longer be viewed as obscure even if
we don't provide a model for them. Particularly, for `dict.__init__`, which has
multiple overloads, it's not even possible to specify a model that'll capture
each override properly.

I'd argue that it is more consistent with what people expect to not have empty
models interfere with the obscurity of a given callable, so let's separate the
SkipOverrides and model concepts out.

We accomplish this by checking if a model is completely empty - in this case,
we simply don't register the model at all, and the existing machinery takes
care of piping things through and marking things as obscure.

Reviewed By: dkgi

Differential Revision: D24009705

fbshipit-source-id: 0f827ba92ad28bd282410a7247566cc861a8e01c