support cross-repository anchor annotations
Summary:
This is part 2 of supporting cross-repository annotations - Pysa can
also be a producer of cross-repository taint. The terminology we use for the
leaves is "anchors", so let's re-use that name in the parsing as well. The rest
of the leaf_name, etc. were designed to be agnostic to what we call the leaves,
so only work that needs to be done is parsing =)

Reviewed By: fahndrich

Differential Revision: D24023433

fbshipit-source-id: df546623241c0c9526e956f8b37b183b9f358681