implement a taint analysis to find missing flows due to missing type information
Summary:
This patch adds another analysis to find missing flows due to missing type information at call sites. It works by using a symbolic callee with a sink `UnknownCallee` on all parameters whenever the analysis could not find callees at a call site. It can be triggered with the parameter `--find-missing-flows=type`. Then, we can use the `leaves order_by="number_issues"` command in `sapp explore` to post-process the results and get a list of locations where we lose the most flows.

Note that the analysis to find missing flows through obscure models is available with `--find-missing-flows=obscure`.

Reviewed By: sinancepel

Differential Revision: D24147899

fbshipit-source-id: 5e59777ac7cfe2a66e79a934aab2695b9d5b012f