start verifying global annotations as well when parsing models
Summary:
Context: Previously, we didn't validate globals at all. We relied on the callable annotation checking to make sure the object existed in the type system. This hid issues where an attribute was

a) Defined in a superclass
b) Corresponded to a property or a callable

Let's start actually validating globals and catching these errors.

Reviewed By: dkgi

Differential Revision: D24169103

fbshipit-source-id: 1c0721c061033b1eac7531ba7328a8592c881096