scaffolding for separate pass to compute call graph information
Summary:
Context: The way Pysa gets its callable information is a bit limited, fundamentally. We've created a hook in type checking called the "CallGraphBuilder" that registers the callees as we see them, and we separately also track the namesof callables in the type system by abusing the typing.Callable type to include the name of callees. This has worked decently well, especially since now we derive the callees from the built call graph, making the discrepancies minimal, but the situation is not perfect.

Particularly, this is a huge problem with decorated functions and callable classes. With decorated functions, we often have cases where the type is something akin to `typing.Callable[..., T] -> CallableClass[T]`, and Pyre drops the name of the undecorated function completely at this point. Similarly, if someone calls a callable class directly, we currently hack in a solution to look for the class name, but can't solve the general case.

To help with these problems, this diff creates the scaffolding for a separate pass that builds the call graph for a given define, and provides a really simple proof-of-concept case - when the type of a provided function is typing.Callable, handle it properly. In the future, we'll extend this to cover all the different cases that Pyre handles.

To be able to reference the type check resolution information outside type checking, we re-use the fixpoint module with a trivial lattice to have consistent keys be provided for callables.

Reviewed By: dkgi

Differential Revision: D24268456

fbshipit-source-id: 61407ed32261bf447753d2ca5a5b632d2c5baa16