Add optional `transitive` flag to MethodsOfSubclassesGenerator
Summary: `get_all_subclass_defines_from_pyre` has a `transitive` flag that recurses through all subclasses (and subclasses of subclasses etc) when true, and only direct subclasses of the base class when false, but it is not currently used by `MethodsOfSubclassesGenerator`. This adds a new optional flag (defaulting to `False` as before) that makes it possible to mark this as `True` so we can catch e.g. `AsyncActionHandler` which is a subclass of a subclass of `AsyncComponent`.

Reviewed By: dkgi

Differential Revision: D24287974

fbshipit-source-id: ef6e94f2c50a56684eafd73e5a54ce59a2bb88b3