support call graph for attributes of optionals
Summary:
One of the cases Pysa diverges from Pyre is in how we treat attribute accesses for optionals. Since Pyre doesn't know the type, it defaults to any. For Pysa, we still care about the callee, so we try to retain it.

The reason we're moving things from call resolution -> call graph is that longer term callResolution.ml will be removed.

Reviewed By: dkgi

Differential Revision: D24294174

fbshipit-source-id: 67e686e98dd31c6f3f8bbe860360a7408e627fe4