Use transforms in kinds
Summary:
We now emit transform kinds instead of the underlying base kind. This required adjusting all the code in processing that traverses traces in order to actually transform the kinds, using the strategy outlined here: https://fburl.com/x8ty73nd

There were 3 places where we had to adjust kinds as we traversed traces:
1. in model generation when tracing from issues to source and sink leaves
2. in trimmed_trace_graph, when traversing from affected frames to issues (this transforms by adding the transform), and again when traversing from issues down to leaves (transforms by stripping the transform)
3. When propagating source kinds along sink traces with --progagate-source-kinds-to-sinks in order to find the correct sink traces.

One tricky aspect is that the issue source/sink kinds should always have the normalized T: form, rather than T@, even if the initial frame contains the transform. We do that in issue processing. The @ should be purely internal to a frame just so we know how it transforms stuff.

During processing, we cache the map (incoming kind, outgoing kind) on trace frames based on the @ kinds associated with the frame.

Differential Revision: D24005997

fbshipit-source-id: e9d38176c2cdbeecc34af1bad0ddf00c06629e91