move get_indirect_target logic to callGraph.ml
Summary:
This diff moves the "resolve overrides, then calculate what a callable type *really* corresponds to* logic to the callGraph module. This gives us the precise override/method tracking in the call graph representation, which was missing from the Pyre-provided call graph by default.

Note: To make this decision, we need the override data, so the tests now record override data in shared memory :(

Reviewed By: dkgi

Differential Revision: D24312124

fbshipit-source-id: f407aaca4b491100ae317e5b78ef08ecbb4dcdeb