Do not crash when `site.getsitepackages()` is not available in virtualenv
Summary:
It turns out that there's a very annoying virtualenv bug: `site.getsitepackages()` is not always available on all python releases and all platforms. See github issue #326. The sad downstream impact for us is that we are affected by it even in cases where we actually don't need to invoke this function.

This diff tries to work around that issue by catching the `AttributeError` and fall back to empty site root with a warning printed.

Reviewed By: pradeep90

Differential Revision: D24365599

fbshipit-source-id: 05363cc2440c8437e2e50652fadd4ee897b6a4a5