start threading `callee kind` when resolving callees and properly resolve staticmethods
Summary:
The basic goal here is to support static methods - regular methods will be bound methods, but if you
access a method through the class or invoke a static method, you won't have a bound method. In this case, we
need to do a bit more work to tag things as methods (unfortunately, Pysa cares about the distinction historically :( ).

Reviewed By: dkgi

Differential Revision: D24344740

fbshipit-source-id: c86961f6850129527bf16aa4dbb67b77d09d5783