make decision on whether to collapse tito in call graph building
Summary:
Context: We used to make the decision of whether to collapse tito inline in forwardAnalysis. This decision requires computing the type of the callee, and is a natural fit for computation here.

(This also lets us unit test the component a lot better.)

For now, the code is duplicated, but will be removed once we integrate this component with the existing stuff.

Reviewed By: dkgi

Differential Revision: D24433532

fbshipit-source-id: 4fc6fe6535245af777657ab81672131ea1b3b66e