support higher order functions
Summary:
A bit of explanation is warranted here: In forward and backward analysis, when we encounter `hof(f, arg)`, we simulate it the following way:

```
$result = f(arg)
hof($result, arg)
```

Therefore, to get correct results, we need to have callee informations for every position. This representation allows us to know which index to exclude from the call to `f` as well as the targets at each point.

The limitation of at most 1 call target unfortunately already exists. It's not clear at all what to do when 2 or more lambdas are passed in.

Reviewed By: dkgi

Differential Revision: D24431593

fbshipit-source-id: f35a02a7d39dbfa583b660956c54d762a2742f7b