Fix glob handling attempt for migration script
Summary:
Debugging upgrade script behavior when multiple configurations are being created and we attempt/fail to use glob targets on some of them.

Restructure the logic to check on validity of globs without committing and having to call a revert or recursively call self.run, which can have unexpected side effects like splitting the commit in two.

Reviewed By: pradeep90

Differential Revision: D24480995

fbshipit-source-id: d7183b04c7abf136f10728b5973bf61915f502ff