Clean up target codemod error suppression
Summary:
Ensure when we revert back from globs, we aren't applying errors to the files we're choosing to omit from the generated configuration.

Additionally, clean up repeated code that lints and checks for errors and just shell out to either the strict default codemod, which can take over, or to the prepackaged method that does this.

Reviewed By: grievejia

Differential Revision: D24554239

fbshipit-source-id: b46d7a7ce5a37036aa06db9c60c8ef228286f316