Consolidate handling of partial error suppression
Summary:
Remove special casing for partial error suppression exceptions; all other subcommands of the error suppressing commands that might encounter this already just let the exception bubble to user.

Include guidance on how to avoid using --unsafe in base mesage.

This will also prevent codemods that are handling multiple configurations to have a partial wipe of progress due to a partial suppression error and continue on to commit an incomplete result. Failing the whole job if `--unsafe` was not passed in is more appropriate behavior.

Reviewed By: MaggieMoss

Differential Revision: D24554238

fbshipit-source-id: 5835d3257e304f4c081becf6d157b4681de7b12a