handle duplicate entries for locations
Summary:
The core problem's captured in the test case being added. Our analysis generates synthetic expressions that share locations with the expression they're defined from in the CFG.

This diff provides a method to support the full complexity of what's happening - it provides a way of handling these synthetic callees by extending our map to be location -> key -> callees instead of location -> callees, and provides a context-free way of heuristically determining name (for when we hit a duplicate).

Future work will optimize representation a bit so that we instead only store names for the colliding cases, but this diff should serve as a bit of documentation to the goal.

Reviewed By: dkgi

Differential Revision: D24552898

fbshipit-source-id: 5c5e41f82777b76c5e041161e514b31fe40e8343