refactor record_toplevel_definition to use of Callable.create
Summary: Context: I will be modifying the call graph building to use the new CallGraph module. We need callable handles for those to work, which we get from defines via the `Callable.create` function. Let's make it clear that there is no separate mapping of define -> callable for the staticAnalysis service.

Reviewed By: dkgi

Differential Revision: D24597450

fbshipit-source-id: 0a770cfdd9fdbebd96e0b64c9ca606a98374a402