Modify ConstructorInitializedAttributeSourceGenerator to validate models
Summary:
The existing model generator finds all the arguments in the class constructor and assumes that they are class attributes with the same name, and creates a taint model for that attribute. However, this can result in a lot of invalid models if the constructor takes arguments that are not set as class attributes.

This diff modifies the model generator to take a `PyreConnection` and query it for class attributes, and only generate models for the constructor parameters that are actually class attributes.

Because querying Pyre takes a long time, the queries are batched (200 seems like a decent number that speeds things up significantly) rather than querying individually for each class.

Reviewed By: gbleaney

Differential Revision: D24561212

fbshipit-source-id: a2e1c4f539cc2e62d7a1969548fea419dad18f20