add support for `any_parameter` constraints
Summary:
Supports the same set of constraints as return annotations for now.

Docs will be updated all together after this stack of diffs is accepted (I want to leave the room open for discussion of syntax first).

Note that we generally need this in Python due to the existence of named + positional parameters, so we can't do the tricks we do in languages like Hack where you have your queries be AnyOf(parameters[0].annotation.blah, parameters[1].annotation.blah).

Reviewed By: dkgi

Differential Revision: D24801351

fbshipit-source-id: f8fc5df01893adc03ee90e345019063ca72cc22b