change mode representation to deduplicate sanitizers
Summary: The basic idea of this diff is that we want to unify the representation of sanitizers, and not have possible entries such as `Sanitize [SanitizeTITO, SanitizeAll, SanitizeSinks]` be allowed in the type. This change models sanitization as a domain, and sets us up to build source- and sink- specific sanitizers.

Reviewed By: dkgi

Differential Revision: D24848710

fbshipit-source-id: 3907cc6b74feec1c7e6592a996834e6035bb69c0