Resolve generic aliases.
Summary:
Problem: We used to resolve primitive aliases, but not generic aliases.

Earlier, if `Foo = Union[T, str]`, we would replace `Foo[int]` with `Foo[Any]`.

The main change is to resolve aliases in parametric types. This requires that we avoid replacing unbound parameters with `Any` in `AttributeResolution.check_invalid_type_parameters`. Added a flag for that and threaded it through the consumers of aliases. Mentioned the different edge cases in a comment.

Added tests for type variable misuses, such as `TypeVar("T", bound="SomeGenericClass")`, which break unless we preserve the implicit behavior that `bound=SomeGenericClass` actually works like `bound=SomeGenericClass[Any]`.

Reviewed By: shannonzhu

Differential Revision: D24529354

fbshipit-source-id: 21e64adf0031aed1db0772b6269adf96085922de