stop requiring extra parameters for models
Summary:
For a long time, Pysa's model parser has properly adjusted parameter positions based on the actual Pyre signature (this is what powers the ability to omit optional parameters in the model syntax). In recent discussions, we have come to the conclusion that also allowing users to omit parameters that are not tainted is a major usability improvement.

This gives us the ability to write things such as

```
Sanitize
def function_with_50_parameters(): ...
```

and

```
def function_with_50_parameters(dangerously_loggable: TaintSink[Logging]) -> TaintSource[Test]: ...
```

Has the side benefit of making the model syntax a lot more consistent to write - you no longer have to worry about specifying a default value for parameters.

Reviewed By: arthaud

Differential Revision: D24966612

fbshipit-source-id: c5200c56a6563b7829642bb7f713b9bcd5b4427b