Add models for wsgiref web server
Summary:
[`wsgiref`](https://docs.python.org/3.8/library/wsgiref.html#module-wsgiref) is a Web Server Gateway Interface (WSGI) reference implementation in the Python standard library. It's built on top of `http.server` so additional models are only needed for the functionality that it adds on top of that library.

To write this diff, I read through [all of the docs on `wsgiref`](https://docs.python.org/3.8/library/wsgiref.html#module-wsgiref) and tainted anything that looked to match the existing categories we have.

This implementation is pretty simple and for the most part does not need model generators for coverage. One place where we could add a model generator for future work is to taint the `environ` argument to the `app` function that is passed into [`make_server`](https://docs.python.org/3.8/library/wsgiref.html#wsgiref.simple_server.make_server). The `app` function is the main entrypoint for handling requests, and it's `environ` argument is a grab bag of headers. Currently, we will not track this argument as tainted, however, if users follow the same pattern as the [example](https://docs.python.org/3.8/library/wsgiref.html#examples) and use the helper functions in [`wsgiref.util`](https://docs.python.org/3.8/library/wsgiref.html#module-wsgiref.util) to process the `environ`, then we will detect the output of those helper functions as tainted.

Reviewed By: dkgi

Differential Revision: D25066921

fbshipit-source-id: 5ac255c74bcaf866fa788166a63b33e87bb353cd