misc flask taint models
Summary:
# request.get_json
`flask.request.get_json` is [inherited](https://github.com/pallets/flask/blob/06495e791b59461cb8cf8c0bb3221f7fd4a14da8/src/flask/wrappers.py#L20) from `flask.wrappers.JSONMixin`, so I've modeled `flask.wrappers.JSONMixin`. It returns the request data as json.
# flask.jsonify
`flask.jsonify` creates a json response from the args and kwargs. I've modeled it as returned to user.
# flask.render_template
This function takes a template name and the context kwargs to be injected into the named template. I haven't tainted the template name because an attacker could not achieve XXS or SSTI as far as I'm aware with this parameter. The worst they could is cause a different template to be rendered. I've modeled the context kwargs as ReturnedToUser, XSS, and SSTI as all of these things can be achieved by controlling them.
# flask.render_template_string
This is the same story as `render_template` except instead of providing the name for a template you provide a string source for the template. The template source string, if controlled, can be an XXS and SSTI sink as well.

Reviewed By: gbleaney

Differential Revision: D25033687

fbshipit-source-id: a3f84da431cc10e72f6616b53ce4ea4a7c48608d