don't verify decorators for model queries
Summary:
When modelling user-specified methods and functions, we require only some decorators (mainly property and x.setter) to exist, in order to distinguish between
propertys, methods and property setters. With the methods that are getting passed in programatically, this is not a problem, so let's not worry about validating the extra decorators.

Note that this approach is subtly different than setting decorators = [] before passing the define into the "create_annotations" method, as that would result in us losing this distinction.

Reviewed By: pradeep90

Differential Revision: D25223498

fbshipit-source-id: 32905267efe979f504a5a288b6206d188b442501