Invalidate cache file when cached type environment doesn't load
Summary: There's a bug that occurs occasionally where it seems to fail to load the type environment from the cache, but then go on to load the callables and overrides from the file successfully. This leads to some very strange results, so let's just delete the cache file if the cached type environment fails to load.

Reviewed By: sinancepel

Differential Revision: D25374761

fbshipit-source-id: 0f82e60b40cf9cc12080d5bd62d472d2e506024e