Ensure that errors are fully specified at the `ModelParser.create` boundary
Summary:
This change brings with it a lot of noise, but the core idea is simple. Previously, we'd construct part of an error message when we threw an InvalidModel error, and the rest would be filled in when this got caught and re-thrown (with callable name, path, location information). This diff makes it so that instead of having string errors, we start having verification errors at the create boundary.

Some knock-on changes had to be done to make this a no-op from the consumer perspective: We push the complexity of "Invalid model for blah" to the verification_error data structure (this is implemented in a non-ideal way right now; it will be addressed in the future when we separate the kind of an error from its location/path/callable name), and we had to modify a lot of code to no longer throw, but instead return result types, and deal with the fallout.

Reviewed By: dkgi

Differential Revision: D25513178

fbshipit-source-id: f36c87446365d833a700ec6b2f68b5267e70a782