make error handling for model verification errors in signatures explicit
Summary:
Same as before - moves us towards a model where in the future, we'll be able to treat errors as objects more easily.

For now, simply pushes the exception throwing outside of modelVerifier.ml and into the model parser. Obviously, we don't want to be throwing
any exceptions here longer term.

Reviewed By: arthaud

Differential Revision: D25500168

fbshipit-source-id: 76f0e52132b662fc74cf33d9d2eb69a60490c4b8