start displaying model verification errors in stdout when verifying models
Summary: This is the milestone we were shooting for - instead of having opaque error messages, we now display structured errors on stdout when Pysa models fail to validate. The next steps once these changes make it to production will be to amend Pyre's client to display these in a nicer way, create a new command which wraps `pyre query "validate_taint_models ()"` to use these JSONs and use a consistent display.

Reviewed By: grievejia

Differential Revision: D25573615

fbshipit-source-id: 278d665702215991ce72508c7c439364e57833fa