Fix broken link in transcluded file
Summary:
If you link to a relative url, e.g `[link](../some-page)`, then docusaurus doesn't change that link at all, so whenever it ends up on a final page, it will be relative to that page, so the destination will differ, depending on what file imports this one.

If you link to a **file** instead, e.g. `[link](../some-page.md)`, then it will be resolved at build time, and remain constant.

So in general, you should be linking to markdown files, rather than url paths.

Reviewed By: gbleaney

Differential Revision: D25587678

fbshipit-source-id: 7cd58f53fa68be37986b7d202927430a6f5c1429