Fix of_approximation
Summary: OverUnderSet domain was inconsistent in how it handled of_approximation. The order of elements mattered w.r.t. duplicates elements (one being in the under approximation). This cleans up the semantics so that `add` means add point-wise and if any element is in the under-approximation then that is reflected.

Reviewed By: sinancepel

Differential Revision: D25594367

fbshipit-source-id: 2196eda47c4c4081f822376457f3e73dfeb1d315