Remove caching of callgraph and overrides and add integration test
Summary:
Computing the callgraph and overrides uses shared memory and causes some odd behavior, resulting in model changes not showing up when artifacts are loaded from cache. This reverts it and goes back to caching only the type environment and initial callables so caching can at least be functional again.

Also added an integration test to ensure that model changes are properly reflected.

Reviewed By: sinancepel

Differential Revision: D25467125

fbshipit-source-id: 57db987de624b21637f822acc99606667befc384