simplify generator handling in backwards analysis
Summary:
We were hand-rolling logic that can be better captured by the generator assignments, already provided to us by the Pyre API. Simplifies how generators are handled, and fixes correctness issues for tuples while doing so.

This also revealed a bug in how we handled d.__iter__().__next__() for dictionary keys, so fix that as well. Note that I chose to merge the handling of `__iter__().__next__()` as they don't occur in isolation, and doing this allows us to get the type of the dictionary more naturally.

Reviewed By: arthaud

Differential Revision: D25727880

fbshipit-source-id: ed6e2b0000a1d7f0aa32494f9d31cda83747292a