Initial implementation of sandbox (#361)
Summary:
This is a basic version of a sandbox service for Pyre that blindly returns type checking results for input provided through `POST` or `GET`. I've set this up on an AWS EC2 instance.

Pull Request resolved: https://github.com/facebook/pyre-check/pull/361

Test Plan: Followed guide in the README.md and checked that the instance returns type checking results.

Reviewed By: MaggieMoss

Differential Revision: D25706608

Pulled By: dkgi

fbshipit-source-id: f62fd01dfcbc22c65e666d12c2c3ab0b54826ba0