Log run id to pyre statistics tables
Summary:
Start logging a consistent run ID to Pyre statistics scuba tables.

The current table logs a single line item per filepath, with various annotation statistics attached. The only way to get an aggregate view of an entire project is to gather many line items, and we currently do so by grouping all line items that have the same "date" - aka a floor of the timestamp.

This winds up being problematic if we have more than one run happen in a calendar day (all stats are doubled). If we group by run ID instead of date, we can easily avoid this. Similarly, we sometimes need to get the latest stat for a particular set of files. Currently we filter data by the last 24 hours, but this can result in edge cases where no data is given or double data is given - would like to group by a path/runID combo, then sort by ID and take the latest. (null run IDs can probably be backfilled with flooring the timestamp to avoid breakage.)

Open to suggestions for any cleaner ways to resolve above concerns, too.

Reviewed By: grievejia

Differential Revision: D25900545

fbshipit-source-id: a17f410f4aaf0fd3ef2bb21eb2d0fa512ea53f79