support calls with named parameter with `ViaValueOf` and `ViaTypeOf` features
Summary:
The current implementation of `ViaValueOf` and `ViaTypeOf` stores the position of the argument to add the value or type of. This is wrong when handling calls with named arguments.
For instance:
```
def f(a, b) -> TaintSource[Test, ViaValueOf[b]]: ...
f(b=2, a=1)
```
This would incorrectly generate a taint with `via-value: 1`. This patch fixes the issue by storing both the name and the position, as a root access path.

Reviewed By: sinancepel

Differential Revision: D25987704

fbshipit-source-id: 3be273decd5fd4d970a4c74932b9823fc418dc80